+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+### v1.0.3
+* Fix: correctly position relic popup if there are multiple pages of relics.
+
 ### v1.0.2
 * Fix: selling relics no longer buys the card behind the click
 * Fix: work around basemod hack that made selling potions seem impossible